requirements.txt
pywin32~=310.0
pycryptodome~=3.22.0
cryptography~=44.0.2
PyPDF2
endesive